binary search algorithm of order o(log n)
Binary search is an algorithm that is based on finding a certain element in a vector based on dividing the length of that array by 2, in addition to adding the desired number to the last index of the vector. With this, we will have an algorithm whose complexity is logarithmic.